New requirement for directly setting user password
Added `2.34` to capture the requirement that a user be able to choose their own password and directly set it in the application without needing to reveal it to another party. The objective of this requirement is to avoid situations where users are given a password that has been set by another party (e.g. administrator) and are not given the ability to change it.