Draft: Add redundancy to 0x24-V19-Config.md
This is a first draft for a rule that would add a redundancy requirement.
High availability is usually a requirement for critical systems as those stated in the Level 3 example for the Healthcare industry (in the old 3.0.1 documents at least). Therefore I restricted this suggested rule to only Level 3.
I'm not sure if there is any corresponding CVSSv3 or CWE reference for keeping a redundant system or failover procedure in place, so I left those out.

With regard to the actual wording I'm open for suggestions. This could either be focusing on the redundancy or the failover part or it could include both. I've just started this draft with the very simple verison of just requiring at least one level of redundancy for all mission critical components. 
Redundancy for all components is certainly nice to have, but not realistically feasible nor is it helping the seucrity of an application.

I think this is security related, because a non-available app is insecure by definition for those cases where that very high security level 3 is required.