Draft: Add backups to 0x24-V19-Config.md
I think this is security relevant because an app that is not available is by definition insecure.
Also the ASVS already includes guidelines for things that should not be relevant during "normal" operation like not including plaintext passwords in the sourcecode. (which users should never have access to anyway - obviously except for open source projects, but there are other rules not applicable to open source as well)

I listed this under 19.5 because it's a more generic concern. Not sure if this would fit under 19.1 or if it's worth establishing an own subheading for this.

Regarding which parts should be covered by a backup, I'd love to hear feedback. Not sure if this sentence I wrote there is getting too convoluted by listing the things to backup. I'm not really happy with the "other custom components", but couldn't think of a better description at this very moment. Maybe a native speaker has a better idea for wording this?

Last but not least: It's super important to not only perform backups, but to also check if you can restore from a backup. That's probably the biggest problem with backups: Even those people who perform regular backups barely ever check if they actually include all the necessary things and if there is a proper process for restoring from a backup. It's too late to check that when you NEED to restore from a backup.