Remove the notion of physical tamper or epoxy
Hello everyone,

I would suggest we change the statement 20.18 to say:

* Verify the presence of physical tamper resistance and/or tamper detection features, including epoxy.
into:
* Verify the presence of tamper resistance and/or tamper detection features.

Tamper evidence and tamper resistance are something like security by obscurity, is it useful, yes, should it be used? (yes in addition to other controls and in very specific use-cases)

Here are some arguments why i changed this:
1) If we look at a run off the mill IoT device, what's the possible most likely threat model?
 - Attacker buys X devices, reverse engineers them to find juicy stuff (shared keys, certs, vulnerable SW or vulns in SW, etc) - Tamper evidence or resistance is moot here. Attacker will shave off epoxy, x-ray the device, etc, etc, and won't care for tamper detection. 
 - Attacker has access to a device, for let's say x hours, but needs to return that device to the owner (think interdiction attacks, covert physical or black bag attacks) - This is a specific issue where tamper evidence is welcome, but in the real world it is not really useful or will be useful. Epoxy is a *extremely* dangerous notion, because people think it's the solution to prevent tampering instead of implementing proper secure boot, trusted execution and encrypted storage on modern SoCs devices.

Epoxy as tamper evidence can be easily removed, and even more so, re-applied (in the case of an interdiction attack), so as a tamper-evident measure, it is of dubious value. If you want tamper detection it needs to be implemented like in modern HSMs or POS terminals, that means you implement something like a security mesh around the SoC / Memory, where it will be a physical-hardware method that will wipe the keys from the device or brick the device as a method to prevent extraction or tampering. But for modern SoCs that is not needed since you already have a secure enclave that if used right, makes those attacks very unfeasible.  

Regarding that, i would also consider upping this requirement to L3. Reason - Proper tamper detection is hard, you should implement trusted execution before that (I'll also send some suggestions in pull requests to other points regarding that). In general it is needed only for most critical stuff (PoS terminals, Medical, etc gear)...

Sincerely,
Tonimir