Proposed rewording of 2.1.8
Verify the use of weak authenticators--such as SMS and email--is limited to secondary verification and transaction approval and not as a replacement to more secure authentication methods. If offering weak authenticators, verify that stronger methods are made available first, that users are aware of the risks, or that proper measures are in place to limit the risks of account compromise.