Password length proposed change
I know that 5.1.1.2 requests at least 8 characters, and i know that by proper use of password hashing (scrypt, argon etc) with salting that the issue of short passwords can be of reduced risk, but people are people. Can we at least insist on something that's not breakable trivially? Today with modern GPUs anything with 8 characters can be bruteforced in under a day (https://twitter.com/hashcat/status/1095807014079512579) can we at least bump this up 50% to 12 characters?

Sincerely,
Tonimir