Resolve #536
Thank you for the very detailed review. Moved several of these controls to V1 architecture as that's where they truly belong, and made them L2. The only one where I didn't take this approach is the client side artifacts that are visible in V14.2.1. I've added a note about that in the explanation.