0x01 - replace "great deal" phrase by "much"
As discussed in #547 "a great deal more", "a great deal better" etc. isn't a well-known english phrase and as many people will read the standard that aren't native speakers, this might be confusing.   
Better stick to a simpler term to express the comparative here. I'd suggest using much as a replacement.